Flip #include inheritance direction
Make TMC5272.h include TMC5272_SPI.h. Previously this was backwards, so it went:
    main <-- TMC5272_SPI <-- TMC5272

Which I did not like.

+++ TMC5272_SPI.h
@@ -7,7 +7,7 @@ This is the implementation of these functions.
 
 */
 
-#include "TMC5272.h"
+#include <stddef.h>
 #include "spi.h"
 #include "mxc_device.h"
 #include "mxc_pins.h"
@@ -19,5 +19,5 @@ This is the implementation of these functions.
 #define SPI_MASTER_SS_INDEX 1      // SS0 = P1_23
 
 /* Function Prototypes */
-void tmc5272_readWriteSPI(uint16_t icID, uint8_t* tx_data, size_t dataLength, uint8_t* rx_data);
+void tmc5272_readWriteSPI(uint16_t icID, uint8_t* tx_data, uint8_t dataLength, uint8_t* rx_data);
 uint8_t tmc5272_getNodeAddress(uint16_t icID);
